fix(gdwfc): cast size_t to int64_t for UtilityFunctions::print on libc++
On macOS (Apple Clang + libc++) size_t is unsigned long, which does not
convert unambiguously to godot::Variant — the variadic print() template
fails to build with an ambiguous conversion error inside
utility_functions.hpp. Linux/GCC happens to pick a conversion that
compiles, but the cast is the correct portable fix.

Other call sites in gdwfc_v2.cpp already cast to int for the same reason.

+++ src/gdwfc.cpp
@@ -159,7 +159,7 @@ PackedInt32Array GDTilingWFC::run() {
             }
 
             UtilityFunctions::print("WFC succeeded! Output size: ", output_width, "x", output_height);
-            UtilityFunctions::print("  Total neighbor rules provided: ", wfc_neighbors.size());
+            UtilityFunctions::print("  Total neighbor rules provided: ", static_cast<int64_t>(wfc_neighbors.size()));
         } else {
             UtilityFunctions::print("WFC failed - contradiction encountered");
         }
